pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "password-cracking-lab-defense-system"
version = "0.1.0"
description = "Password Cracking Lab + Defense System with ML/DL based anomaly detection and response."
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.32.0",
  "sqlalchemy>=2.0.36",
  "alembic>=1.13.3",
  "psycopg[binary]>=3.2.3",
  "pydantic>=2.9.2",
  "pydantic-settings>=2.6.1",
  "email-validator>=2.2.0",
  "python-jose[cryptography]>=3.3.0",
  "passlib[argon2]>=1.7.4",
  "prometheus-client>=0.21.0",
  "opentelemetry-distro>=0.52b1",
  "opentelemetry-exporter-otlp>=1.27.0",
  "opentelemetry-instrumentation-fastapi>=0.52b1",
  "opentelemetry-instrumentation-sqlalchemy>=0.52b1",
  "mlflow>=2.17.2",
  "aiokafka>=0.11.0",
  "scikit-learn>=1.5.2",
  "numpy>=2.1.3",
  "pandas>=2.2.3",
  "torch>=2.4.1",
  "joblib>=1.4.2",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.3",
  "httpx>=0.27.2",
  "ruff>=0.7.1",
  "locust>=2.32.4",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
